feat: moved all svg logic into Icon package
