+++ packages/core/styles/src/types/index.ts
@@ -82,19 +82,4 @@ export type GradientDirection =
   | "to bottom left"
   | "to bottom right";
 
-/**
- * Available gradient variants for components.
- * Currently supports the Nimbus AI gradient styling.
- */
-export type GradientTypes = "linear";
-
 export type GradientColorStop = string; // Flexible string to support "red", "red 50%", "rgba(255,0,0,0.5)", etc.
-
-/**
- * Represents a color stop in a gradient with offset position and color value.
- * Used for parsing and creating SVG gradient definitions.
- */
-export type GradientStop = {
-  offset: string;
-  color: string;
-};
